Functional Requirement 3
The remaining five cards are randomly
generated. Three cards with random colors and values between 1 to 6, with an additional sign, that is either a plus (+) or a minus (-). The remaining two cards have an 80 percent chance of being a signed card. One card can be a flip (+/-) card. One card can be a double (x2) card. Flip cards and double cards do not have a color, they only change the values or
the sign of the last played card.