Update Test.java
Program asks user to stand or keep on and finishes the set if user or computer board sum is equal 20 ,or if in boards, there is 9 card. Program analyzes the cards in the boards if the cards all are blue, the game will finish.Program asks user to use any card from their hand.